+++ src/token.rs
@@ -66,6 +66,7 @@ pub enum TipoToken {
     // Palavras-chave de orientação a objetos (Fase 2)
     Classe,
     Registro,
+    Enum,
     Herda,
     Metodo,
     Construtor,
@@ -168,6 +169,7 @@ pub fn palavra_chave(texto: &str) -> Option<TipoToken> {
         "nulo" => TipoToken::Nulo,
         "classe" => TipoToken::Classe,
         "registro" => TipoToken::Registro,
+        "enum" => TipoToken::Enum,
         "herda" => TipoToken::Herda,
         "metodo" => TipoToken::Metodo,
         "construtor" => TipoToken::Construtor,